+++ src/utils/constants.ts
@@ -21,3 +21,11 @@ export const enum USER_PERMS {
   DeleteList = 'lists.item.delete',
   ExportList = 'lists.item.export.get'
 }
+
+export const RELATED_USERS_URL = 'lists/related-users';
+export const RELATED_USERS_LIMIT = 1000;
+
+export const enum RELATED_USERS_TYPE {
+  CreatedBy = 'create',
+  UpdatedBy = 'update'
+}